+++ Write-ProgressEx/Write-ProgressEx.psm1
@@ -1,4 +1,4 @@
-﻿# mazzy@mazzy.ru, 2018-05-06
+﻿# mazzy@mazzy.ru, 2018-05-13
 # https://github.com/mazzy-ax/Write-ProgressEx
 
 #region Module variables
@@ -9,7 +9,7 @@ $ProgressExDefault = @{
     MessageOnFirstIteration = {param([hashtable]$pInfo) Write-Warning "[$(Get-Date)] Id=$($pInfo.Id):$($pInfo.Activity):$($pInfo.Status): start."}
     MessageOnNewActivity    = {param([hashtable]$pInfo) Write-Warning "[$(Get-Date)] Id=$($pInfo.Id):$($pInfo.Activity):$($pInfo.Status):"}
     MessageOnNewStatus      = {param([hashtable]$pInfo) Write-Warning "[$(Get-Date)] Id=$($pInfo.Id):$($pInfo.Activity):$($pInfo.Status):"}
-    MessageOnCompleted      = {param([hashtable]$pInfo) Write-Warning "[$(Get-Date)] Id=$($pInfo.Id):$($pInfo.Activity):$($pInfo.Status): done. Iterations=$($pInfo.Current), Elapsed=$($pInfo.stopwatch.Elapsed)"}
+    MessageOnCompleted      = {param([hashtable]$pInfo) Write-Warning "[$(Get-Date)] Id=$($pInfo.Id):$($pInfo.Activity):$($pInfo.Status): done. Iterations=$($pInfo.Current), Elapsed=$($pInfo.Elapsed))"}
 }
 
 $StdParmNames = (Get-Command Write-Progress).Parameters.Keys
@@ -62,7 +62,11 @@ function Get-ProgressEx {
             $pInfo.Clone()
         }
         elseif ( $Force ) {
-            @{Id = $id; Reset = $true}
+            @{
+                Id             = $id
+                Reset          = $true
+                UpdateInterval = [TimeSpan]::FromMilliseconds(100)
+            }
         }
         else {
             $null
@@ -88,24 +92,15 @@ function Write-ProgressExMessage {
     .NOTES
     This function is not exported
     #>
-    [cmdletbinding(DefaultParameterSetName = 'CustomMessage')]
+    [cmdletbinding()]
     param(
         [Parameter(Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
         [hashtable]$pInfo,
 
-        [Parameter(Position = 1, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'CustomMessage')]
         [scriptblock[]]$Message,
-
-        [Parameter(ParameterSetName = 'StdMessage')]
         [switch]$ShowMessagesOnFirstIteration,
-
-        [Parameter(ParameterSetName = 'StdMessage')]
         [switch]$ShowMessagesOnNewActivity,
-
-        [Parameter(ParameterSetName = 'StdMessage')]
         [switch]$ShowMessagesOnNewStatus,
-
-        [Parameter(ParameterSetName = 'StdMessage')]
         [switch]$ShowMessagesOnCompleted
     )
 
@@ -136,11 +131,15 @@ function Set-ProgressEx {
     Set parameters for the progress with Id and dispaly this values to the console.
 
     .DESCRIPTION
-    The cmdlet:
-    * save parameters
-    * display this parameters to console
+    The cmdlet performs a technical work based on user parameters in the pInfo:
+
+    * save parameters to global structure
+    * display this parameters to console via standard Write-Progress
+    * show event messages
+    * decorate activity with iteration info
     * complete progress if Completed parameter is $true
-    * complete all children progresses always.
+    * complete all children progresses always
+    * and so on
 
     .EXAMPLE
     $range = 1..1000
@@ -174,30 +173,34 @@ function Set-ProgressEx {
             $pInfo.Completed = $true
         }
 
-        Write-ProgressExMessage @{
-            pInfo                        = $pInfo
-            ShowMessagesOnFirstIteration = $pInfo.Reset
-            ShowMessagesOnNewActivity    = $pInfo.Activity -ne $ProgressEx[$pInfo.Id].Activity
-            ShowMessagesOnNewStatus      = $pInfo.Status -ne $ProgressEx[$pInfo.Id].Status
-            ShowMessagesOnCompleted      = $pInfo.Completed
-        }
+        Write-ProgressExMessage $pInfo `
+            -ShowMessagesOnFirstIteration:($pInfo.Reset) `
+            -ShowMessagesOnNewActivity:($pInfo.Activity -ne $ProgressEx[$pInfo.Id].Activity) `
+            -ShowMessagesOnNewStatus:($pInfo.Status -ne $ProgressEx[$pInfo.Id].Status) `
+            -ShowMessagesOnCompleted:($pInfo.Completed)
 
-        if ( $pInfo.Completed ) {
-            $pInfo.Stopwatch = $null
-            $ProgressEx.Remove($pInfo.Id)
-        }
-        else {
-            $pInfo.Reset = $false
-            $ProgressEx[$pInfo.Id] = $pInfo
+        $shouldShowProgressBar = switch ( $pInfo.ShowProgressBar ) {
+            'None' { $false }
+            'Force' { $true }
+            Default {
+                try {
+                    $pInfo.Completed -or
+                    $pInfo.UpdateInterval -le (New-TimeSpan -End $pInfo.ProgressDateTime -Start $pInfo.UpdateDateTime)
+                }
+                catch {
+                    $true
+                }
+            }
         }
 
         # Invoke standard write-progress cmdlet
-        if ( -not $pInfo.NoProgressBar ) {
+        if ( $shouldShowProgressBar ) {
             $pArgs = @{}
             $pInfo.Keys | Where-Object { $StdParmNames -contains $_ } | ForEach-Object {
                 $pArgs[$_] = $pInfo[$_]
             }
 
+            # Decorate activity
             if ( $pInfo.Total ) {
                 $pArgs.Activity = $pArgs.Activity, ($pInfo.Current, $pInfo.Total -join '/') -join ': '
             }
@@ -211,6 +214,16 @@ function Set-ProgressEx {
             }
 
             Write-Progress @pArgs
+
+            $pInfo.UpdateDateTime = $pInfo.ProgressDateTime
+        }
+
+        if ( $pInfo.Completed ) {
+            $ProgressEx.Remove($pInfo.Id)
+        }
+        else {
+            $pInfo.Reset = $false
+            $ProgressEx[$pInfo.Id] = $pInfo
         }
 
         # Recursive complete own children
@@ -223,6 +236,17 @@ function Write-ProgressEx {
     <#
     .SYNOPSIS
     Powershell Extended Write-Progress cmdlet.
+
+    .DESCRIPTION
+    The cmdlet Write-ProgressEx contains a 'business-logic':
+
+    * PercentComplete, SecondsRemaining calculation and other calculations,
+    * increment and autoincrement for a Current,
+    * autoname an activity string
+    * smart parentId search,
+    * and so on
+
+    This cmdlet calls Set-ProgressEx to perform technical work and to call a standard Write-Progress.
 
     .EXAMPLE
     Write-ProgressEx -Total $nodes.Count
@@ -246,25 +270,15 @@ function Write-ProgressEx {
     }
     Write-ProgressEx -complete
 
-    .EXAMPLE
-    Ideal: is it possible?
-
-    $nodes | Where-Object ...... | Write-ProgressEx | ForEach-Object {
-        $names | Where-Object ...... | Write-ProgressEx -id 1 | ForEach-Object {
-            ......
-        }
-    }
-    write-posProgress -complete
-
-    .NOTE
+    .NOTES
     Commands 'Write-ProgressEx.ps1' and 'Write-ProgressEx -Complete' are equivalents.
     The cmdlet complete all children progresses.
 
-    .NOTE
+    .NOTES
     A developer can use a parameter splatting.
     See Get-ProgressEx example.
 
-    .NOTE
+    .NOTES
     Cmdlet is not safe with multi-thread.
 
     #>
@@ -296,10 +310,16 @@ function Write-ProgressEx {
 
         [switch]$Reset,
         [switch]$Increment,
-        [System.Diagnostics.Stopwatch]$Stopwatch,
 
-        # The Сmdlet does not call a standard write-progress cmdlet. Thus the progress bar does not show.
-        [switch]$NoProgressBar,
+        [DateTime]$ProgressDateTime,
+        [DateTime]$StartDateTime,
+        [TimeSpan]$Elapsed,
+
+        [DateTime]$UpdateDateTime,
+        [TimeSpan]$UpdateInterval,
+
+        [ValidateSet('Auto', 'Force', 'None')] # Different first letters make typing fast
+        [string]$ShowProgressBar,
 
         # Message templates
         [scriptblock[]]$MessageOnFirstIteration,
@@ -325,29 +345,42 @@ function Write-ProgressEx {
                 $pInfo[$_] = $PSBoundParameters[$_]
             }
 
+            if ( -not $ProgressDateTime ) {
+                $pInfo.ProgressDateTime = Get-Date
+            }
+
             # auto parentId
             if ( $pInfo.Reset -and $pInfo.Keys -notcontains 'ParentId' ) {
                 $ParentProbe = $ProgressEx.Keys | Where-Object { $_ -lt $pInfo.id } | Measure-Object -Maximum
-                $pInfo.ParentId = if ( $null -ne $ParentProbe.Maximum ) { $ParentProbe.Maximum } else { -1 }
+                if ( $null -ne $ParentProbe.Maximum ) {
+                    $pInfo.ParentId = $ParentProbe.Maximum
+                }
             }
 
-            if ( $pInfo.Reset ) {
-                $pInfo.PercentComplete = 0
+            if ( $pInfo.Reset -and -not $Current ) {
                 $pInfo.Current = 0
-                $pInfo.stopwatch = [System.Diagnostics.Stopwatch]::StartNew()
             }
 
-            if ( $pInfo.Increment -or $isPipe ) {
+            if ( $pInfo.Reset -and -not $StartDateTime ) {
+                $pInfo.StartDateTime = $pInfo.ProgressDateTime
+            }
+
+            if ( -not $Elapsed -and $pInfo.StartDateTime ) {
+                $pInfo.Elapsed = New-TimeSpan -End $pInfo.ProgressDateTime -Start $pInfo.StartDateTime
+            }
+
+            $pInfo.Increment = $Increment -or $isPipe
+            if ( $pInfo.Increment ) {
                 $pInfo.Current += 1
             }
 
-            if ( ($pInfo.Total -gt 0) -and ($pInfo.Current -gt 0) -and ($pInfo.Current -le $pInfo.Total) ) {
+            if ( $pInfo.Total -and $pInfo.Current ) {
                 if ( -not $PercentComplete ) {
-                    $pInfo.PercentComplete = [Math]::Min( [Math]::Max(0, [int]($pInfo.Current / $pInfo.Total * 100)), 100)
+                    [int]$pInfo.PercentComplete = [Math]::Min([Math]::Max(0, $pInfo.Current / $pInfo.Total * 100), 100)
                 }
 
-                if ( -not $SecondsRemaining -and $pInfo.Stopwatch ) {
-                    $pInfo.SecondsRemaining = [Math]::Max(0, 1 + $pInfo.stopwatch.Elapsed.TotalSeconds * [Math]::Max(0, $pInfo.Total - $pInfo.Current) / $pInfo.Current)
+                if ( -not $SecondsRemaining -and $pInfo.Elapsed ) {
+                    [int]$pInfo.SecondsRemaining = [Math]::Max(0, $pInfo.Elapsed.TotalSeconds * ($pInfo.Total - $pInfo.Current) / $pInfo.Current)
                 }
             }